+++ src/mod-compatibility/all.ts
@@ -1,2 +1,3 @@
 import './nax-art-switch'
 import './lqm-joern-mod'
+import './jetpack'