+++ src/mod-compatibility/jetpack.ts
@@ -1,0 +1,10 @@
+export {}
+declare global {
+    namespace ig {
+        namespace Input {
+            interface KnownActions {
+                'keys-jump': true
+            }
+        }
+    }
+}